separated products and productOptions tests
